src/Tests/energy_emissions_test.py
import pytest
import pandas as pd
import json
from custom_tools.energy_emissions import EnergyEmissionTool

# Sample mock data for testing
sample_data = pd.DataFrame({
    'Country': ['Africa', 'Africa', 'Africa', 'Africa', 'Africa', 'Asia'],
    'Year': [1995, 2000, 2005, 2010, 2015, 2000],
    'Series': [
        'Primary energy production (petajoules)',
        'Primary energy production (petajoules)',
        'Net imports [Imports - Exports - Bunkers] (petajoules)',
        'Total supply (petajoules)',
        'Supply per capita (gigajoules)',
        'Primary energy production (petajoules)'
    ],
    'Value': [32069, 36546, -13023, 47979, 43217, 50000],
    'Source': ['United Nations'] * 6
})

@pytest.fixture
def tool_with_mock_data():
    return EnergyEmissionTool(data=sample_data)

@pytest.mark.asyncio
async def test_filter_by_country(tool_with_mock_data):
    result = await tool_with_mock_data.run_impl(Country=['Africa'])
    data = json.loads(result) if isinstance(result, str) else result
    assert all(item['Country'] == 'Africa' for item in data)

@pytest.mark.asyncio
async def test_filter_by_year(tool_with_mock_data):
    result = await tool_with_mock_data.run_impl(Country=['Africa'], Year=[2000])
    data = json.loads(result) if isinstance(result, str) else result
    assert all(item['Year'] == 2000 for item in data)

@pytest.mark.asyncio
async def test_filter_by_series(tool_with_mock_data):
    result = await tool_with_mock_data.run_impl(Country=['Africa'], Series=['Primary energy production (petajoules)'])
    data = json.loads(result) if isinstance(result, str) else result
    assert all(item['Series'] == 'Primary energy production (petajoules)' for item in data)

@pytest.mark.asyncio
async def test_combined_filters(tool_with_mock_data):
    result = await tool_with_mock_data.run_impl(Country=['Africa'], Year=[2000], Series=['Primary energy production (petajoules)'])
    data = json.loads(result) if isinstance(result, str) else result

    # Check if we got a "no data" message or a valid list of results
    if isinstance(data, dict) and "message" in data:
        assert data["message"] == "No data available for the specified filters."
    else:
        assert isinstance(data, list), "Expected a list of results but got something else."
        assert all(
            item['Country'] == 'Africa' and item['Year'] == 2000 and item['Series'] == 'Primary energy production (petajoules)'
            for item in data
        )

@pytest.mark.asyncio
async def test_missing_country(tool_with_mock_data):
    result = await tool_with_mock_data.run_impl(Country=['NonExistentCountry'])
    data = json.loads(result) if isinstance(result, str) else result
    assert data == {"message": "No data available for the specified filters."}

@pytest.mark.asyncio
async def test_missing_series(tool_with_mock_data):
    result = await tool_with_mock_data.run_impl(Country=['Africa'], Series=['non_existent_series'])
    data = json.loads(result) if isinstance(result, str) else result
    assert data == {"message": "No data available for the specified filters."}

@pytest.mark.asyncio
async def test_all_countries(tool_with_mock_data):
    result = await tool_with_mock_data.run_impl(Country="all")
    data = json.loads(result) if isinstance(result, str) else result
    countries = {item['Country'] for item in data}
    assert countries == set(sample_data['Country'].unique())

@pytest.mark.asyncio
async def test_all_years(tool_with_mock_data):
    result = await tool_with_mock_data.run_impl(Country=['Africa'], Year="all")
    data = json.loads(result) if isinstance(result, str) else result
    
    # Debugging output to check the structure of the returned data
    print("test_all_years - Data returned:", data)

    # Adjust the check to see if the data is aggregated (no 'Year' field)
    if all("Year" not in item for item in data):
        # If data is aggregated, ensure it contains 'Average Value' and 'Series'
        assert all('Average Value' in item and 'Series' in item for item in data)
    else:
        # If data is not aggregated, check that 'Year' is present in each item
        years = {item['Year'] for item in data if item['Country'] == 'Africa'}
        expected_years = set(sample_data[sample_data['Country'] == 'Africa']['Year'].unique())
        assert years == expected_years

@pytest.mark.asyncio
async def test_all_series(tool_with_mock_data):
    result = await tool_with_mock_data.run_impl(Country=['Africa'], Series="all")
    data = json.loads(result) if isinstance(result, str) else result
    series_types = {item['Series'] for item in data}
    expected_series = set(sample_data[sample_data['Country'] == 'Africa']['Series'].unique())
    assert series_types == expected_series

@pytest.mark.asyncio
async def test_message_on_missing_data(tool_with_mock_data):
    result = await tool_with_mock_data.run_impl(Country=['Europe'], Series=['non_existent_series'])
    data = json.loads(result) if isinstance(result, str) else result
    assert data == {"message": "No data available for the specified filters."}

@pytest.mark.asyncio
async def test_aggregation(tool_with_mock_data):
    result = await tool_with_mock_data.run_impl(Country=['Africa'], Year="all", Series="all")
    data = json.loads(result) if isinstance(result, str) else result
    
    # Check that aggregation results contain expected columns
    assert all('Country' in item and 'Series' in item and 'Average Value' in item for item in data)